feat: root 프로젝트 설정. api(ui) 패키지 모듈 분리
