+++ api/build.gradle
@@ -1,0 +1,7 @@
+project(':api') {
+	dependencies {
+		implementation 'org.springframework.boot:spring-boot-starter-web'
+		testImplementation 'org.springframework.boot:spring-boot-starter-test'
+		implementation project(':service')
+	}
+}
