+++ api/src/test/java/kitchenpos/domain/api/TableRestControllerTest.java
@@ -1,12 +1,12 @@
-package kitchenpos.table.ui;
+package kitchenpos.domain.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kitchenpos.table.application.TableService;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.dto.OrderTableRequest;
-import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.table.dto.OrderTableUpdateEmptyRequest;
-import kitchenpos.table.dto.OrderTableUpdateNumberOfGuestsRequest;
+import kitchenpos.api.TableRestController;
+import kitchenpos.service.table.TableService;
+import kitchenpos.service.table.dto.OrderTableRequest;
+import kitchenpos.service.table.dto.OrderTableResponse;
+import kitchenpos.service.table.dto.OrderTableUpdateEmptyRequest;
+import kitchenpos.service.table.dto.OrderTableUpdateNumberOfGuestsRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -43,8 +43,7 @@ class TableRestControllerTest {
     @Test
     void test_get() throws Exception {
         //given
-        given(tableService.list()).willReturn(Collections.singletonList(new OrderTableResponse(OrderTable.of(1,
-                true))));
+        given(tableService.list()).willReturn(Collections.singletonList(new OrderTableResponse()));
 
         //then
         mockMvc.perform(get("/api/tables"))
@@ -55,7 +54,7 @@ class TableRestControllerTest {
     @Test
     void test_post() throws Exception {
         //given
-        given(tableService.create(any())).willReturn(new OrderTableResponse(OrderTable.of(1, true)));
+        given(tableService.create(any())).willReturn(new OrderTableResponse());
 
         //then
         mockMvc.perform(post("/api/tables").content(objectMapper.writeValueAsString(new OrderTableRequest()))
@@ -67,7 +66,7 @@ class TableRestControllerTest {
     @Test
     void test_put_changeEmpty() throws Exception {
         //given
-        given(tableService.changeEmpty(any(), any())).willReturn(new OrderTableResponse(OrderTable.of(1, true)));
+        given(tableService.changeEmpty(any(), any())).willReturn(new OrderTableResponse());
 
         //then
         mockMvc.perform(put("/api/tables/{orderTableId}/empty", 0).content(objectMapper.writeValueAsString(new OrderTableUpdateEmptyRequest()))
@@ -79,7 +78,7 @@ class TableRestControllerTest {
     @Test
     void test_put_changeNumberOfGuests() throws Exception {
         //given
-        given(tableService.changeNumberOfGuests(any(), any())).willReturn(new OrderTableResponse(OrderTable.of(1, true)));
+        given(tableService.changeNumberOfGuests(any(), any())).willReturn(new OrderTableResponse());
 
         //then
         mockMvc.perform(put("/api/tables/{orderTableId}/number-of-guests", 0).content(objectMapper.writeValueAsString(new OrderTableUpdateNumberOfGuestsRequest()))
