+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'kitchenpos'
+include 'api'
+include 'service'
+include 'domain'
